requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
dnspython==2.4.2
faiss-cpu==1.7.4
fastapi==0.95.2
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
langchain==0.0.186
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.1.1
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
websocket-client==1.6.4
websockets==11.0.3
yarl==1.9.2
